Unification of ring_allocator usages
